SNC-4: Code refactored.
