+++ app/src/main/java/com/stanislavdumchykov/socialnetworkclient/data/UserStore.kt
@@ -6,7 +6,7 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
-import com.stanislavdumchykov.socialnetworkclient.domain.utils.Constants
+import com.stanislavdumchykov.socialnetworkclient.presentation.utils.Constants
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
