+++ app/src/main/java/com/stanislavdumchykov/socialnetworkclient/data/di/DataStoreModule.kt
@@ -8,7 +8,7 @@ import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.emptyPreferences
 import androidx.datastore.preferences.preferencesDataStoreFile
-import com.stanislavdumchykov.socialnetworkclient.domain.utils.Constants
+import com.stanislavdumchykov.socialnetworkclient.presentation.utils.Constants
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
