+++ app/src/main/java/com/stanislavdumchykov/socialnetworkclient/domain/repository/UsersRepository.kt
@@ -3,5 +3,5 @@ package com.stanislavdumchykov.socialnetworkclient.domain.repository
 import com.stanislavdumchykov.socialnetworkclient.domain.model.User
 
 interface UsersRepository {
-    suspend fun getHardcodedUsers(): List<User>
+    suspend fun getUsers(): List<User>
 }
