--- app/src/main/java/com/stanislavdumchykov/socialnetworkclient/domain/utils/ScreenList.kt
@@ -1,5 +1,0 @@
-package com.stanislavdumchykov.socialnetworkclient.domain.utils
-
-enum class ScreenList {
-    MY_PROFILE, CONTACT_LIST
-}
