+++ app/src/main/java/com/stanislavdumchykov/socialnetworkclient/presentation/utils/Constants.kt
@@ -1,12 +1,10 @@
-package com.stanislavdumchykov.socialnetworkclient.domain.utils
+package com.stanislavdumchykov.socialnetworkclient.presentation.utils
 
 object Constants {
     const val PERCENT_33 = 0.33f
     const val PERCENT_50 = 0.5f
     const val PERCENT_60 = 0.6f
     const val PERCENT_100 = 1f
-    const val PASSWORD_MIN_LENGTH = 8
-    const val PASSWORD_PATTERN = "\\d+|\\w+"
     const val PREFERENCES_EMAIL = "EMAIL"
     const val PREFERENCES_PASSWORD = "PASSWORD"
     const val USER_PREFERENCES_NAME = "user_preferences"
