+++ app/src/main/java/com/stanislavdumchykov/socialnetworkclient/presentation/utils/NavigationRoutes.kt
@@ -1,0 +1,7 @@
+package com.stanislavdumchykov.socialnetworkclient.presentation.utils
+
+enum class NavigationRoutes {
+    SignUp,
+    Pager,
+    ContactProfile
+}
