+++ app/src/main/java/com/stanislavdumchykov/socialnetworkclient/presentation/utils/ScreenList.kt
@@ -1,0 +1,5 @@
+package com.stanislavdumchykov.socialnetworkclient.presentation.utils
+
+enum class ScreenList {
+    MY_PROFILE, CONTACT_LIST
+}
